+++ src/include/globals.h
